Added favicon and site header icons and updated documents section
I've added the Haiku Inc favicon, recreated in Inkscape to look like the original. I've added this to the site header as well. I've also updated the Documents section, adding a full PDF with all the pages for the IRS Certificate of Incorporation as well as fixing the broken link leading to the IRS filings.